+++ tools/osx/drive/NuxeoFinderSync/FinderSync.swift
@@ -31,7 +31,7 @@ class FinderSync: FIFinderSync {
                                                           name: self.watchFolderNotif,
                                                           object: nil)
 
-        let triggerURL = URL(string: "nxdrive://trigger_watch")
+        let triggerURL = URL(string: "nxdrive://trigger-watch")
         NSWorkspace.shared.open(triggerURL!)
 
         // Set up images for our badge identifiers.
@@ -117,7 +117,7 @@ class FinderSync: FIFinderSync {
 
         // Access online
         let item1 = NSMenuItem(title: "Access online",
-                               action: #selector(openInBrowser(_:)),
+                               action: #selector(accessOnline(_:)),
                                keyEquivalent: "A")
         item1.image = self.icon
         menu.addItem(item1)
@@ -129,15 +129,22 @@ class FinderSync: FIFinderSync {
         item2.image = self.icon
         menu.addItem(item2)
 
+        // Edit metadata
+        let item3 = NSMenuItem(title: "Edit metadata",
+                               action: #selector(editMetadata(_:)),
+                               keyEquivalent: "E")
+        item3.image = self.icon
+        menu.addItem(item3)
+
         return menu
     }
 
-    @IBAction func openInBrowser(_ sender: AnyObject?) {
+    @IBAction func accessOnline(_ sender: AnyObject?) {
         // Event fired by "Access online" menu entry
         let items = FIFinderSyncController.default().selectedItemURLs()
         for item in items! {
-            NSLog("openInBrowser: target: %@", item.path as NSString)
-            openNXUrl(command: "access", target: item)
+            NSLog("accessOnline: target: %@", item.path as NSString)
+            openNXUrl(command: "access-online", target: item)
         }
     }
 
@@ -146,7 +153,16 @@ class FinderSync: FIFinderSync {
         let items = FIFinderSyncController.default().selectedItemURLs()
         for item in items! {
             NSLog("copyShareLink: target: %@", item.path as NSString)
-            openNXUrl(command: "share_link", target: item)
+            openNXUrl(command: "copy-share-link", target: item)
+        }
+    }
+
+    @IBAction func editMetadata(_ sender: AnyObject?) {
+        // Event fired by "Edit metadata" menu entry
+        let items = FIFinderSyncController.default().selectedItemURLs()
+        for item in items! {
+            NSLog("editMetadata: target: %@", item.path as NSString)
+            openNXUrl(command: "edit-metadata", target: item)
         }
     }